v1.0: Modify unique constraint in `db_model.py`
The unique constraint in `smon` table of the database model is modified. The constraint previously was encompassing four attributes, i.e., name, port, http, and body but now it has been simplified to just cover name and port. This change was made to better reflect the unique entries in the database.