v1.0.4: Improve session and group validation logic
This commit enhances the session and group validation in the application. Specifically, the system now sets a default value of 5 for the session TTL if no valid value is provided. In addition, it improves the group id validation by comparing the group id in cookies with the group id obtained from database. Lastly, it refactors how user roles and groups are fetched from the database.