v1.1: Update optional strings to default as None and improve error handling
The main change in this commit is the update of optional strings in class_models.py to default as None instead of an empty string for better data consistency and less confusion. In addition to this, enhancements were made to the error handling logic in script.js to handle 401 error status in Ajax requests gracefully. Condition for displaying features based on user role in new_receiver.html was also updated and the function name of buttons in channels.html changed for better readability.